requirements.txt
streamlit
gradio_client
pillow
opencv-python
langgraph
matplotlib
google-generativeai
